+++ Changes.md
@@ -12,7 +12,7 @@ Features
 Improvements
 ------------
 
-- USD : Added automatic render-time translation of UsdPreviewSurface shaders to Arnold.
+- USD : Added automatic render-time translation of UsdPreviewSurface shaders and UsdLuxLights for Arnold.
 - SceneReader :
   - Added support for Alembic visibility attributes.
   - Added loading of `treatAsPoint` and `treatAsLine` parameters from UsdLux lights.